requirements.txt
torch>=1.8.0
numpy>=1.19.0
scipy>=1.6.0
scikit-learn>=0.24.0
pandas>=1.2.0
networkx>=2.5
PyYAML>=5.4.0
tqdm>=4.60.0
matplotlib>=3.3.0
seaborn>=0.11.0